requirements.txt
pandas
requests
matplotlib
isodate
openpyxl
google-api-python-client
python-dotenv
